# swagger.yaml
openapi: 3.0.0
info:
  title: MindLoo API
  version: 2.0.0
  description: |
    Complete RESTful API for managing patients, therapists, appointments, and therapy-related operations.
    
    ## Features
    - Patient and Therapist Management
    - Appointment Scheduling
    - Therapy Session Tracking
    - Medical Records Management
    - Secure Messaging System
    - Review and Rating System
    - Treatment Planning and Goal Tracking
    
  contact:
    name: Stefan Ramač
    url: https://www.linkedin.com/in/stefanramac
servers:
  - url: http://localhost:3000
    description: Development server
...
components:
  schemas:
...
    Therapist:
      type: object
      required:
        - therapistId
        - firstName
        - lastName
        - email
      properties:
        therapistId:
          type: string
          example: "T001"
        firstName:
          type: string
          example: "Jane"
        lastName:
          type: string
          example: "Smith"
        email:
          type: string
          format: email
          example: "jane.smith@mindloo.com"
        type:
          type: string
          default: "therapist"
        profile:
          type: object
          properties:
            age:
              type: integer
              example: 35
            gender:
              type: string
              example: "female"
            specialization:
              type: string
              example: "Cognitive Behavioral Therapy"
            experience:
              type: integer
              example: 10
        contactInfo:
          type: object
          properties:
            phone:
              type: string
            address:
              type: string
            place:
              type: string
            country:
              type: string
    
    # Appointment Schema
...
    Review:
      type: object
      required:
        - therapistId
        - patientId
        - rating
      properties:
        reviewId:
          type: string
          readOnly: true
        therapistId:
          type: string
        patientId:
          type: string
        appointmentId:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        categories:
          type: object
          properties:
            professionalism:
              type: integer
              minimum: 1
              maximum: 5
            communication:
              type: integer
              minimum: 1
              maximum: 5
            effectiveness:
              type: integer
              minimum: 1
              maximum: 5
            empathy:
              type: integer
              minimum: 1
              maximum: 5
        comment:
          type: string
        isAnonymous:
          type: boolean
          default: false
        isVerified:
          type: boolean
          default: false
    
    # Treatment Plan Schema
...
paths:
...
  /api/therapists:
    post:
      tags:
        - Therapists
      summary: Create a new therapist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Therapist'
      responses:
        '201':
          description: Therapist created successfully
        '400':
          description: Therapist already exists
    
    get:
      tags:
        - Therapists
      summary: Get all therapists
      responses:
        '200':
          description: List of therapists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Therapist'
  /api/therapists/{id}:
    get:
      tags:
        - Therapists
      summary: Get therapist by ID or email
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Therapist data
        '404':
          description: Therapist not found
    
    patch:
      tags:
        - Therapists
      summary: Update therapist information
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Therapist'
      responses:
        '200':
          description: Therapist updated successfully
    
    delete:
      tags:
        - Therapists
      summary: Delete a therapist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Therapist deleted successfully
  
  # Appointments Endpoints
...
  /api/reviews/therapist/{therapistId}:
    get:
      tags:
        - Reviews
      summary: Get all reviews for a therapist
      parameters:
        - name: therapistId
          in: path
          required: true
          schema:
            type: string
        - name: isVisible
          in: query
          schema:
            type: boolean
        - name: minRating
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 5
      responses:
        '200':
          description: List of reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Review'
...
  /api/treatment-plans/therapist/{therapistId}:
    get:
      tags:
        - Treatment Plans
      summary: Get all treatment plans by a therapist
      parameters:
        - name: therapistId
          in: path
          required: true
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of treatment plans
...